+++ next/src/routes/_layout/questionnaire.$questionnaireId/_layout-q/variables/new.tsx
@@ -17,7 +17,7 @@ export const Route = createFileRoute(
   component: RouteComponent,
   errorComponent: ({ error }) => (
     <CreateVariableLayout>
-      <ErrorComponent error={error.message} />
+      <ErrorComponent error={error} />
     </CreateVariableLayout>
   ),
   loader: async ({