+++ next/src/routes/_layout/questionnaire.$questionnaireId/_layout-q/variables/variable.$variableId.tsx
@@ -16,7 +16,7 @@ export const Route = createFileRoute(
   component: RouteComponent,
   errorComponent: ({ error }) => (
     <EditVariableLayout>
-      <ErrorComponent error={error.message} />
+      <ErrorComponent error={error} />
     </EditVariableLayout>
   ),
   loader: async ({